installer: add & use securelyDeleteItemAtPathIfItExists() where appropriate

+++ tunnelblick/installer.m
@@ -312,6 +312,24 @@ void errorExitIfSymlinksOrDoesNotExistOrIsNotReadableAtPath(NSString * path) {
 
 void securelyDeleteItemAtPath(NSString * path) {
 
+    errorExitIfAnySymlinkInPath(path);
+
+    const char * pathC = [path fileSystemRepresentation];
+
+    if (  0 != unlink(pathC)  ) {
+        appendLog([NSString stringWithFormat: @"unlink() failed with error %d ('%s') for path %@", errno, strerror(errno), path]);
+        errorExit();
+    }
+}
+
+void securelyDeleteItemAtPathIfItExists(NSString * path) {
+
+    if (  ! [gFileMgr fileExistsAtPath: path]  ) {
+        return;
+    }
+
+    errorExitIfAnySymlinkInPath(path);
+
     const char * pathC = [path fileSystemRepresentation];
 
     if (  0 != unlink(pathC)  ) {
@@ -516,10 +534,7 @@ void securelyCopyDirectly(NSString * sourcePath, NSString * targetPath) {
         errorExit();
     }
 
-    if (  ! [gFileMgr tbRemovePathIfItExists: targetPath]  ) {
-        appendLog([NSString stringWithFormat: @"Unable to remove %@", targetPath]);
-        errorExit();
-    }
+    securelyDeleteItemAtPathIfItExists(targetPath);
 
     if (  ! securelyCreateFileOrDirectoryEntry(isDir, targetPath)  ) {
         errorExit();
@@ -1085,17 +1100,14 @@ void safeCopyOrMovePathToPath(NSString * sourcePath, NSString * targetPath, BOOL
 	// see the .tblk (which has been copied) but not the config.ovpn (which hasn't been copied yet), so it complains.
 	NSString * dotTempPath = [targetPath stringByAppendingPathExtension: @"temp"];
 	errorExitIfAnySymlinkInPath(dotTempPath);
-	if ( [gFileMgr fileExistsAtPath:dotTempPath]  ) {
-		[gFileMgr tbRemoveFileAtPath:dotTempPath handler: nil];
-	}
-	
+    securelyDeleteItemAtPathIfItExists(dotTempPath);
+
 	createFolder([dotTempPath stringByDeletingLastPathComponent]);
 	
 // JKB: DO securelyCopy HERE, TOO
     if (  ! [gFileMgr tbCopyPath: sourcePath toPath: dotTempPath handler: nil]  ) {
 		appendLog([NSString stringWithFormat: @"Failed to copy %@ to %@", sourcePath, dotTempPath]);
-		[gFileMgr tbRemovePathIfItExists: dotTempPath];
-
+        securelyDeleteItemAtPathIfItExists(dotTempPath);
 		errorExit();
 	}
 	appendLog([NSString stringWithFormat: @"Copied %@\n    to %@", sourcePath, dotTempPath]);
@@ -1466,12 +1478,12 @@ void testSecurelyRename(NSString * folder) {
 
     // Touch two files (delete them first if they exist)
     NSString * test1Path = [folder stringByAppendingPathComponent: @"rename-test-target-1"];
-    [gFileMgr tbRemovePathIfItExists: test1Path];
+    securelyDeleteItemAtPathIfItExists(test1Path);
     if (  ! [gFileMgr createFileAtPath: test1Path contents: nil attributes: nil]  ) {
         appendLog([NSString stringWithFormat: @"testSecurelyRename Can't create rename-test-target-1 in %@", folder]);
     }
     NSString * test2Path = [folder stringByAppendingPathComponent: @"rename-test-target-2"];
-    [gFileMgr tbRemovePathIfItExists: test1Path];
+    securelyDeleteItemAtPathIfItExists(test2Path);
     if (  ! [gFileMgr createFileAtPath: test2Path contents: nil attributes: nil]  ) {
         appendLog([NSString stringWithFormat: @"testSecurelyRename Can't create rename-test-target-2 in %@", folder]);
     }
@@ -1480,8 +1492,9 @@ void testSecurelyRename(NSString * folder) {
     BOOL good = (0 != renamex_np([test1Path fileSystemRepresentation], [test2Path fileSystemRepresentation],(RENAME_NOFOLLOW_ANY | RENAME_EXCL)));
 
     // Delete the files
-    [gFileMgr tbRemovePathIfItExists: test1Path];
-    [gFileMgr tbRemovePathIfItExists: test1Path];
+    securelyDeleteItemAtPathIfItExists(test1Path);
+    securelyDeleteItemAtPathIfItExists(test2Path);
+
 
     if (  ! good  ) {
         appendLog([NSString stringWithFormat: @"Rename test failed for %@", folder]);
@@ -2265,10 +2278,7 @@ void pruneFolderAtPath(NSString * path) {
 		
 		NSDirectoryEnumerator * innerEnum = [gFileMgr enumeratorAtPath: pruneCandidatePath];
 		if (  ! [innerEnum nextObject]  ) {
-			if (  ! [gFileMgr tbRemovePathIfItExists: pruneCandidatePath]  ) {
-				errorExit();
-			}
-			
+            securelyDeleteItemAtPathIfItExists(pruneCandidatePath);
 			appendLog([NSString stringWithFormat: @"Removed folder because it was empty: %@", pruneCandidatePath]);
 		}
 	}
@@ -2282,9 +2292,8 @@ void exportToPath(NSString * exportPath) {
 	
 	// Remove the output file if it already exists
 	// (We do this so user doesn't do something with it before we're finished).
-	if (  ! [gFileMgr tbRemovePathIfItExists: tarPath]  ) {
-		errorExit();
-	}
+    securelyDeleteItemAtPathIfItExists(tarPath);
+    
 	// Create a temporary folder
 	NSString * tempFolderPath = newTemporaryDirectoryPath();
 	
@@ -2659,9 +2668,7 @@ void mergeForcedPreferences(NSString * sourcePath) {
 		}
 		
 		if (  modifiedExistingPreferences  ) {
-			if (  ! [gFileMgr tbRemovePathIfItExists: targetPath]  ) {
-				errorExit();
-			}
+            securelyDeleteItemAtPathIfItExists(targetPath);
 			if (  ! [existingPreferences writeToFile: targetPath atomically: YES]  ) {
 				appendLog([NSString stringWithFormat: @"Error: could not write %@  ", sourcePath]);
 				errorExit();
